--- src/main/java/domain/payment/PaymentCalculator.java
@@ -1,14 +1,0 @@
-package domain.payment;
-
-/**
- * 지불 금액을 계산하는 객체
- *
- * @author 토니
- * @version 1.0
- * <p>
- * 날짜 : 2020/03/05
- */
-public class PaymentCalculator {
-
-// TODO: 2020/03/05 ??
-}
